Theme, Music, Sound, UI Updates + Scene 5 Updates (#6)
* Updated Character Positions

Updated character positions for all encounters

* Changed Resize Setting

Removed ability for player to resize window (causes layout issues depending on the computer resolution)

* Sounds and Music Added

- Laughing woman sound
- New song/music for ambience

* AI Glitching Images Added

Added Al glitching images / containers to the Al program

* Added Immediate Change to Al Program

Added a change to the call with Martin to immediately turn on Program Al after the 3rd meeting

* Added Sound and Image Visibility to Al Program

- Added creepy laugh
- Added visibility and loops for the glitch to occur after the last Al Encounter

* Removed Fade Out from all Timelines

Removed Fade Out from all Timelines

* Definition and Folder Structure updates

Internal changes to Dialogic timelines (variable names and folder structure)

* Added Martin Theme

Added Martin Theme

* Theme Updates

Theme-1682566781 was added for testing but ended up not being used so it was deleted. Folder structure and changes were pushed before the theme was deleted.

* Added Testing Button

Skips all encounters by changing a variable so that meeting with Martin #3 is chosen.

* Added New Icons

Added Icons for all Desktop Icons

* Updated Al's name to Alpha

Changed Al's name in Dialogic to Alpha to remove the name showing up in dialogic timelines (for example, meeting with Martin, everytime he says "Al-" it shows as a different colour

* Added Fourth Meeting with Martin

Added timeline for 4th meeting with Martin (scene where Martin has X's on his eyes and behind him it says "Open Al"

* Added Function to Reduce Lines of Code

Added Function to Reduce Lines of Code to Al Script

* Updated Video Request Script

- Updated click sound audio play as the previous link was broken
- Added section in func _ready() that changes text of call request depending on scene

* Added Fourth Meeting with Martin Features

- Added black screen to desktop page
- Added scripts to link third meeting to fourth meeting
- Added new image for "Family Photo" to be added later
- Updated Dialogic folder structure

* Added New Icons

Added all new icons to desktop

* Updated Third Meeting With Martin

Updated the text Martin says at the end of the third meeting

* Updated Default Theme

Updated Default Theme for testing of if the effects for names shows up (Does nothing to the actual theme)

* defaul_env.tres Shows No Changes

Unsure of why this was changed at all. i've added it to the commit for future review

---------

Co-authored-by: Vanessa Truong <v.truong9710@gmail.com>